service tests moved to one file, backup test fixed
